requirements.txt
python-telegram-bot>=21.0
telethon>=1.36.0
cryptg>=0.4.0
supabase>=2.10.0
boto3>=1.35.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
rapidfuzz>=3.10.0
